Add the signals.py.
