--- doc/uml/class_signal.uml
@@ -1,7 +1,0 @@
-@startuml
-class SignalControl
-class SignalMessage
-class SignalError
-class SignalStatus
-class SignalLimitSwitch
-@enduml
