+++ python/lsst/ts/m2gui/__init__.py
@@ -31,6 +31,13 @@ else:
 
 from .enums import LocalMode, LimitSwitchType, Ring
 from .utils import set_button
+from .signals import (
+    SignalControl,
+    SignalMessage,
+    SignalError,
+    SignalStatus,
+    SignalLimitSwitch,
+)
 from .log_window_handler import LogWindowHandler
 from .fault_manager import FaultManager
 from .model import Model
